+++ samples/pipe_client.cpp
@@ -12,7 +12,7 @@ class MyChannel : public PipeSession{
 		}
 
 		virtual ~MyChannel(){}
-		virtual bool handle_event(knet::NetEvent evt) { dlog("handle net event {}", evt);  return true; }
+		virtual bool handle_event(knet::NetEvent evt) { dlog("handle net event {}", static_cast<uint32_t>(evt));  return true; }
 		virtual int32_t handle_message(const std::string_view& msg,uint64_t obdata = 0 ) {
 			ilog("---------------{}----------------", msg.size()); 
 			ilog("outband data is {}", obdata);
